Validate the COMBINE scalars, and make the RRF constant a real number
Four COMBINE arguments were taken on trust. Measured against a redis:8
reference, three defects:

1. Silent truncation. The shared scalar parser reaches `std::from_chars`,
   which reports success on whatever numeric prefix it finds rather than
   requiring the whole token. So `WINDOW 20abc` became 20, `WINDOW 1.5`
   became 1, `CONSTANT 60abc` became 60, `CONSTANT 1e3` became 1, and
   `WINDOW 0x10` became 0 -- which this engine reads as unlimited. Acting on
   a number the caller did not write is worse than either engine's answer,
   and the reference rejects all of them. CONSTANT, WINDOW, ALPHA and BETA
   now insist the token is entirely the number.

2. The RRF constant was a uint32_t, and the reference's is a real number.
   Fractional constants are meaningful there -- 1 scores 0.7, 1.5 scores
   0.5818, 2 scores 0.5 -- and we truncated 1.5 to 1. It is now a double.
   Zero stays valid, as in the reference; negative is refused, where the
   reference answers infinity.

3. ALPHA and BETA accepted `inf`, `-inf` and `1e400`, each of which reaches
   a fused score and poisons every comparison against it. The reference
   rejects all three.

Defect 3 hid behind a build flag worth recording: the first fix used
`std::isfinite` and changed nothing, because this project compiles with
`-ffast-math`, which implies `-ffinite-math-only` and lets the compiler fold
that call to `true`. The check now tests the IEEE exponent bits, which is
what src/expr/value.cc and src/indexes/scoring/scorer.h already do.

`WINDOW 0` keeps its meaning of "do not cap the arms". The reference rejects
it, so nothing outside this repo pins it, and COMBINE FUNCTION depends on it:
a user expression is expected to see every candidate, so FUNCTION with no
WINDOW defaults to zero rather than inheriting the RRF/LINEAR default of 20.

Tests, none of which existed before:
  * Eleven parser unit tests over the four arguments -- junk, non-finite,
    negative, out-of-range and the values that must still be accepted --
    plus the zero window and the FUNCTION default.
  * Three FT.HYBRID integration tests: the malformed arguments are rejected,
    `WINDOW 0` returns the full union rather than an empty reply, and
    `CONSTANT 1.5` ranks strictly between 1 and 2, which is what catches a
    truncation that a single-value check would not.
  * Compatibility sweeps of the fractional constants and of weights outside
    [0, 1], both of which the reference accepts.

unsupported_tests.md 5.5c records the remaining differences, all of them
error-message or spelling differences that change no answer, and 5.5d records
`WINDOW 0`.

+++ src/query/multi_search.h
@@ -58,7 +58,9 @@ inline constexpr absl::string_view kDefaultOutputScoreName{"__score"};
 struct FusionConfig {
   enum class Method { kRRF, kLinear, kFunction };
   Method method = Method::kRRF;
-  uint32_t rrf_constant = 60;
+  // Fractional, to match the reference: see rank_fusion.h.
+  double rrf_constant = 60.0;
+  // 0 means unlimited, which is also what COMBINE FUNCTION defaults to.
   uint32_t window = 20;
   std::optional<double> alpha;  // required when method == kLinear
   std::optional<double> beta;   // required when method == kLinear
